Use resolve_deferred_bindings as a separate resolveDeferredBindings util
